requirements.txt
python-dotenv
azure-core
azure-ai-documentintelligence
tenacity
azure-search-documents
azure-search
ipython
openai
azure-storage-blob
langchain
tiktoken
streamlit
